Crate pallet_election_provider_multi_block

Multi-phase, multi-block, election provider pallet.

Overall idea

pallet_election_provider_multi_phase provides the basic ability for NPoS solutions to be computed offchain (essentially anywhere) and submitted back to the chain as signed or unsigned transaction, with sensible configurations and fail-safe mechanisms to ensure system safety. Nonetheless, it has a limited capacity in terms of number of voters it can process in a single block.

This pallet takes pallet_election_provider_multi_phase, keeps most of its ideas and core premises, and extends it to support paginated, multi-block operations. The final goal of this pallet is scale linearly with the number of blocks allocated to the elections. Moreover, the amount of work that it does in one block should be bounded and measurable, making it suitable for a parachain. In principle, with large enough blocks (in a dedicated parachain), the number of voters included in the NPoS system can grow significantly (yet, obviously not indefinitely).

Note that this pallet does not consider how the recipient is processing the results. To ensure scalability, of course, the recipient of this pallet’s data (i.e. pallet-staking) must also be capable of pagination and multi-block processing.

Companion pallets

This pallet is essentially hierarchical. This particular one is the top level one. It contains the shared information that all child pallets use. All child pallets depend on the top level pallet ONLY, but not the other way around. For those cases, traits are used.

This pallet will only function in a sensible way if it is peered with its companion pallets.

The verifier pallet provides a standard implementation of the verifier::Verifier. This pallet is mandatory.

The unsigned module provides the implementation of unsigned submission by validators. If this pallet is included, then Config::UnsignedPhase will determine its duration.

The signed module provides the implementation of the signed submission by any account. If this pallet is included, the combined Config::SignedPhase and Config::SignedValidationPhase will determine its duration

Pallet Ordering:

The ordering of these pallets in a runtime should be:

parent

verifier

signed

unsigned

This is critical for the phase transition to work.

This should be manually checked, there is not automated way to test it.

Pagination

Most of the external APIs of this pallet are paginated. All pagination follow a pattern where if N pages exist, the first paginated call is function(N-1) and the last one is function(0). For example, with 3 pages, the elect of [ElectionProvider] is expected to be called as elect(2) -> elect(1) -> elect(0). In essence, calling a paginated function with index 0 is always a signal of termination, meaning that no further calls will follow.

Phases

The timeline of pallet is overall as follows:

 <  Off  >
0 ------- 12 13 14 15 ----------- 20 ---------25 ------- 30
	           |       |              |            |          |
	     Snapshot      Signed   SignedValidation  Unsigned   Elect

Duration of Snapshot is determined by Config::Pages.

Duration of Signed, SignedValidation and Unsigned are determined by Config::SignedPhase, Config::SignedValidationPhase and Config::UnsignedPhase respectively.

Config::Pages calls to elect are expected, but all in all the pallet will close a round once elect(0) is called.

The pallet strives to be ready for the first call to elect, for example elect(2) if 3 pages.

This pallet can be commanded to to be ready sooner with Config::Lookahead.

Given this, it is rather important for the user of this pallet to ensure it always terminates election via elect before requesting a new one.

Feasible Solution (correct solution)

All submissions must undergo a feasibility check. Signed solutions are checked on by one at the end of the signed phase, and the unsigned solutions are checked on the spot. A feasible solution is as follows:

all of the used indices must be correct.

present exactly correct number of winners.

any assignment is checked to match with PagedVoterSnapshot.

the claimed score is valid, based on the fixed point arithmetic accuracy.

Emergency Phase and Fallback

Config::Fallback is called on each page. It typically may decide to:

Do nothing,

Force us into the emergency phase

computer an onchain from the give page of snapshot. Note that this will be sub-optimal, because the proper pagination size of snapshot and fallback will likely differ a lot.

Note that configuring the fallback to be onchain computation is not recommended, unless for test-nets for a number of reasons:

The solution score of fallback is never checked to be match the “minimum” score. That being said, the computation happens onchain so we can trust it.

The onchain fallback runs on the same number of voters and targets that reside on a single page of a snapshot, which will very likely be too much for actual onchain computation. Yet, we don’t have another choice as we cannot request another smaller snapshot from the data provider mid-election without more bookkeeping on the staking side.

If onchain solution is to be seriously considered, an improvement to this pallet should re-request a smaller set of voters from T::DataProvider in a stateless manner.

Signed Phase

Signed phase is when an offchain miner, aka, polkadot-staking-miner should operate upon. See signed for more information.

Unsigned Phase

Unsigned phase is a built-in fallback in which validators may submit a single page election, taking into account only the [ElectionProvider::msp] (most significant page). See crate::unsigned for more information.

Re-exports

pub use pallet::*;

pub use types::*;

Modules

benchmarking

helpers

Some helper functions/macros for this crate.

pallet

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.

signed

The signed pallet The signed phase of the multi-block election system.

types

Common types of the pallet

unsigned

The unsigned pallet

verifier

The verifier pallet

weights

The weight module

Macros

log

Emit a log specific to this pallet, setting the target to crate::LOG_PREFIX

miner_log

Emit a log from within the offchain miner.

sublog

Emit a log within a submodule of the pallet

Structs

Continue

A fallback implementation that silently continues into the next page.

InitiateEmergencyPhase

A fallback implementation that transitions the pallet to the emergency phase.

Enums

AdminOperation

Different operations that the Config::AdminOrigin can perform on the pallet.

ElectionError

Internal errors of the pallet. This is used in the implementation of [ElectionProvider].

Constants

LOG_PREFIX

The common logginv prefix of all pallets in this crate.

Traits

WeightInfo

Weight functions needed for pallet_election_provider_multi_block.