Module pallet_election_provider_multi_block::signed

The signed pallet The signed phase of the multi-block election system.

Signed submissions work on the basis of keeping a queue of submissions from random signed accounts, and sorting them based on the best claimed score to the worse.

Once the time to evaluate the signed phase comes (Phase::SignedValidation), the solutions are checked from best-to-worse claim, and they end up in either of the 3 buckets:

If they are the first, correct solution (and consequently the best one, since we start evaluating from the best claim), they are rewarded.

Any solution after the first correct solution is refunded in an unbiased way.

Any invalid solution that wasted valuable blockchain time gets slashed for their deposit.

Future Plans:

Lazy deletion: Overall, this pallet can avoid the need to delete any storage item, by:

outsource the storage of solution data to some other pallet.

keep it here, but make everything be also a map of the round number, so that we can keep old storage, and it is ONLY EVER removed, when after that round number is over. This can happen for more or less free by the submitter itself, and by anyone else as well, in which case they get a share of the the sum deposit. The share increases as times goes on. Metadata update: imagine you mis-computed your score.

Re-exports

pub use pallet::*;

Modules

pallet

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.

Structs

SubmissionMetadata

All of the (meta) data around a signed submission

SubstrateWeight

Explore all weights Weights for pallet_election_provider_multi_block::signed using the Substrate node and recommended hardware.

Traits

WeightInfo

Explore all weights Weight functions needed for pallet_election_provider_multi_block::signed.