pallet_election_provider_multi_block::unsigned

Module miner

The miner. The miner code for the EPMB pallet.

It is broadly consisted of two main types:

crate::unsigned::miner::BaseMiner, which is more generic, needs parameterization via crate::unsigned::miner::MinerConfig, and can be used by an external implementation.

crate::unsigned::miner::OffchainWorkerMiner, which is more opinionated, and is used by this pallet via the offchain_worker hook to also mine solutions during the Phase::Unsigned.

Structs

BaseMiner

A base miner that is only capable of mining a new solution and checking it against the state of this pallet for feasibility, and trimming its length/weight.

MaxWinnersFinal

Helper type that computes the maximum total winners across all pages.

MineInput

Aggregator for inputs to BaseMiner.

OffchainWorkerMiner

A miner that is suited to work inside offchain worker environment.

Enums

MinerError

The errors related to the BaseMiner.

OffchainMinerError

The errors related to the OffchainWorkerMiner.

SnapshotType

The type of the snapshot.

Traits

MinerConfig

Configurations for the miner.

Type Aliases

FullSupportsOfMiner

The full version of PageSupportsOfMiner.

PageSupportsOfMiner

Parameterized BoundedSupports for the miner.