Module pallet_election_provider_multi_block::verifier

The verifier pallet

The Verifier Pallet

Feasibility Check

Before explaining the pallet itself, it should be explained what a verification even means. Verification of a solution page (crate::unsigned::miner::MinerConfig::Solution) includes the process of checking all of its edges against a snapshot to be correct. For instance, all voters that are presented in a solution page must have actually voted for the winner that they are backing, based on the snapshot kept in the parent pallet.

After checking all of the edges, a handful of other checks are performed:

Check that the total number of winners is sufficient (DesiredTargets).

Check that the claimed score (sp_npos_elections::ElectionScore) is correct,

and more than the minimum score that can be specified via Verifier::set_minimum_score.

Check that all of the bounds of the solution are respected, namely Verifier::MaxBackersPerWinner, Verifier::MaxWinnersPerPage and Verifier::MaxBackersPerWinnerFinal.

Note that the common factor of all of these checks is that they can ONLY be checked after all pages are already verified. So, In the case of a multi-page verification, these checks are performed at the last page.

The errors that can arise while performing the feasibility check are encapsulated in verifier::FeasibilityError.

Modes of Verification

The verifier pallet provide two modes of functionality:

Single-page, synchronous verification. This is useful in the context of single-page, emergency, or unsigned solutions that need to be verified on the fly. This is similar to how the old school multi-phase pallet works.

Multi-page, asynchronous verification. This is useful in the context of multi-page, signed solutions.

Both of this, plus some helper functions, is exposed via the Verifier trait.

Queued Solution

once a solution has been verified, it is called a queued solution. It is sitting in a queue, waiting for either of:

being challenged and potentially replaced by better solution, if any.

being exported as the final outcome of the election.

Re-exports

pub use __tt_error_token_17 as tt_error_token;

pub use __tt_default_parts_23 as tt_default_parts;

pub use __tt_extra_parts_23 as tt_extra_parts;

pub use __tt_default_parts_v2_23 as tt_default_parts_v2;

Modules

benchmarking

dispatchables

Auto-generated docs-only module listing all defined dispatchables for this pallet.

storage_types

Auto-generated docs-only module listing all (public and private) defined storage types for this pallet.

Structs

Pallet

The Pallet struct, the main type that implements traits and standalone functions within the pallet.

QueuedSolution

A wrapper interface for the storage items related to the queued solution.

SubstrateWeight

Weights for pallet_election_provider_multi_block::verifier using the Substrate node and recommended hardware.

Enums

Call

Contains a variant per dispatchable extrinsic that this pallet has.

Event

The Event enum of this pallet

FeasibilityError

Errors that can happen in the feasibility check.

Status

The status of this pallet.

VerificationResult

Simple enum to encapsulate the result of the verification of a candidate solution.

Traits

AsynchronousVerifier

Something that can do the verification asynchronously.

Config

Configuration trait of this pallet.

SolutionDataProvider

Something that can provide candidate solutions to the verifier.

Verifier

The interface of something that can verify solutions for other sub-pallets in the multi-block election pallet-network.

WeightInfo

Weight functions needed for pallet_election_provider_multi_block::verifier.

Functions

feasibility_check_page_inner_with_snapshot

Same as feasibility_check_page_inner, but with a snapshot.

Type Aliases

ModuleDeprecated

Type alias to Pallet, to be used by construct_runtime.